% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate-da.R
\name{annotate_da}
\alias{annotate_da}
\title{Annotate the Reasoning for a Section of Code}
\usage{
annotate_da(.msg, .expr)
}
\arguments{
\item{.msg}{A character string containing the explanatory message for the code.}

\item{.expr}{An expression representing the R code to be executed.}
}
\value{
Invisible NULL. This function is used for its side effect of printing to the console.
}
\description{
This function prints a formatted message explaining the reasoning for a section of code
and then prints the result of evaluating the provided expression. This function is
intended to be run interactively in the R console to aid in code understanding and debugging.
}
\details{
If the \code{mrgda.annotate_da_file} option is set to a file path, the explanation
and printed result are also appended to that file as Markdown.
}
\examples{
\dontrun{
dat <- Theoph
dat$Time[2] <- 0
annotate_da("Subject 1 has duplicate time 0", {
  dat \%>\% filter(Subject == 1, Time == 0)
})

use_annotate_da_file("script/data-assembly/study-101-annotations.md")
annotate_da("Subject 1 has duplicate time 0", {
  dat \%>\% filter(Subject == 1, Time == 0)
})
}

}
