% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign-id.R
\name{assign_id}
\alias{assign_id}
\title{Assign Subject IDs (creates an ID column)}
\usage{
assign_id(.data, .previously_derived_path = NULL, .subject_col = "USUBJID")
}
\arguments{
\item{.data}{A data.frame to have ID column mutated onto.}

\item{.previously_derived_path}{A string containing the file path of the previous derived data set}

\item{.subject_col}{A character string containing the column name of the subject identifying column in the data}
}
\description{
Upon first use, an ID column will be mutated to the given data set according
to the specified subject column.

For each subsequent use, IDs will be assigned to match the previous derived data
set. If new subjects are found in the data, they will be assigned a
unique value not present in the previous derived data (starting at the maximum
of the previous IDs plus 1).
}
\examples{
Theoph2 <-
  Theoph \%>\%
  dplyr::mutate(USUBJID = paste0("Subject-", Subject)) \%>\%
  dplyr::select(-Subject)

df_with_id <-
  assign_id(
    Theoph2,
    .subject_col = "USUBJID"
  )

}
