% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-egfr.R
\name{calc_egfr}
\alias{calc_egfr}
\title{EGFR formula}
\usage{
calc_egfr(.age, .wt, .serum_creatinine, .sex, .female_value)
}
\arguments{
\item{.age}{age in years}

\item{.wt}{weight in kilograms}

\item{.serum_creatinine}{serum creatinine mg/dL}

\item{.sex}{numeric value representing male & female}

\item{.female_value}{numeric representing female subjects}
}
\description{
EGFR formula
}
\keyword{internal}
