% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-id.R
\name{calc_id}
\alias{calc_id}
\title{ID formula}
\usage{
calc_id(.usubjid, .startid = 1)
}
\arguments{
\item{.usubjid}{subject identifier}

\item{.startid}{numerical value to start creating ID's}
}
\description{
ID formula
}
\keyword{internal}
