% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v-helpers.R
\name{check_for_v_pkgs}
\alias{check_for_v_pkgs}
\title{Checks if all packages needed for v() are present}
\usage{
check_for_v_pkgs()
}
\description{
Returns a vector with the missing packages, or returns NULL if all are
present.
}
\keyword{internal}
