% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-src.R
\name{check_src}
\alias{check_src}
\title{Run all source data checks}
\usage{
check_src(.src_list, .subject_col = "USUBJID")
}
\arguments{
\item{.src_list}{List of all soure domain data.frames}

\item{.subject_col}{Column name of subject identifier (character)}
}
\description{
Run all source data checks
}
