% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-src-duplicates.R
\name{check_src_duplicates}
\alias{check_src_duplicates}
\title{Check if duplicates exist in source data domain}
\usage{
check_src_duplicates(
  .domain_df,
  .domain_name,
  .cols_check = NULL,
  .subject_col = "USUBJID"
)
}
\arguments{
\item{.domain_df}{source data (data.frame)}

\item{.domain_name}{Name of source domain (character)}

\item{.cols_check}{Columns to search duplicates across (ie. c("USUBJID", "PCDTC"))}

\item{.subject_col}{Column name of subject identifier (character)}
}
\description{
Checks if any duplicates records exist in a given data.frame. Some commonly used
domains have default columns to use for the duplicates check, however the .cols_check
argument can be used to customize the columns to be used for the duplicate check.
}
\examples{
\dontrun{
check_src_duplicates(
  .domain_df = src_list$dm,
  .domain_name = "dm",
  .cols_check = NULL,
  .subject_col = "USUBJID")
check_src_duplicates(
  .domain_df = src_list$ex,
  .domain_name = "ex",
  .cols_check = c("USUBJID", "EXSTDTC", "EXTRT"))
check_src_duplicates(
  .domain_df = src_list$pc,
  .domain_name = "pc")
}
}
