% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v-helpers.R
\name{check_subject_col}
\alias{check_subject_col}
\title{Search for ID columns across list of dataframes}
\usage{
check_subject_col(.df_list, .id_cols = c("USUBJID", "ID"))
}
\arguments{
\item{.df_list}{A dataframe \emph{\strong{or}} list of dataframes that you want to process and view.}

\item{.id_cols}{vector of id columns to search for}
}
\value{
a character string
}
\description{
Search for the specified ID columns across list of dataframes, and return the column that appears the most.
Returns \code{NULL} if none are found.
}
\keyword{internal}
