% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v-helpers.R
\name{create_v_datatable}
\alias{create_v_datatable}
\title{Creates a formatted \code{DT::datatable}}
\usage{
create_v_datatable(
  .df,
  .subject_col = NULL,
  .freeze_cols = NULL,
  .digits = 3,
  .show_filters = FALSE,
  .show_labels = TRUE,
  .wrap_labels = FALSE,
  .trunc_labels = FALSE,
  .trunc_length = 20,
  .ft_size = 9,
  .scroll_y = 400,
  .subj_contrast = FALSE
)
}
\arguments{
\item{.df}{A dataframe that you want to process and view.}

\item{.subject_col}{A character string specifying the subject column. If this column is
present in the dataframe, it will be used to group rows in the datatable.}

\item{.freeze_cols}{A character vector specifying columns to freeze when scrolling horizontally.
\code{.subject_col} will automatically be appended to this list (i.e. \code{.subject_col} is always frozen).}

\item{.digits}{number of digits to round numeric columns to. Set to \code{NULL} to prevent rounding.}

\item{.show_filters}{Logical (\code{TRUE}/\code{FALSE}). If \code{TRUE}, show column filters.}

\item{.show_labels}{Logical (\code{TRUE}/\code{FALSE}). If \code{TRUE}, show label attributes associated with the columns.}

\item{.wrap_labels}{Logical (\code{TRUE}/\code{FALSE}). If \code{TRUE}, wrap column labels to have multuple lines.}

\item{.trunc_labels}{Logical (\code{TRUE}/\code{FALSE}). If \code{TRUE}, truncate the column label attributes.}

\item{.trunc_length}{Numeric. Maximum width of column labels. Only relevant if \code{.trunc_labels = TRUE}.}

\item{.ft_size}{Numeric. Base font size of the table. Other labels scale off this value.}

\item{.scroll_y}{Numeric. Height of the table in pixels. Passed to \code{scrollY} option of \code{DT::datatable}.}

\item{.subj_contrast}{Logical (\code{TRUE}/\code{FALSE}). If \code{TRUE}, increase the color contrast between unique \code{.subject_col} groupings.}
}
\value{
An instance of \code{DT::datatable} class with the processed dataframe.
}
\description{
Creates a formatted \code{DT::datatable} with many adjustable formatting options. Designed to be called within
\code{v()} or \code{v_shiny_internal} for easier formatting adjustments and server-side processing.
}
\details{
If \code{.subject_col} is \code{NULL}, the column names \code{"USUBJID"} and \code{"ID"} will be searched and set if present.
If \code{.subject_col} is present in the dataframe, it will be frozen and alternate in color.

Columns with less than 20 unique values are converted to factors. This can be helpful for filtering, which
will show up as the shiny equivalent of \code{selectInput}, as opposed to the default \code{textInput} for character columns.

Columns with label attributes will be appended to column headers as a new line.

You can \emph{drag and drop} columns to move them around in the table. If a column is dragged to a frozen column's location,
the new column will be frozen instead.
}
\keyword{internal}
