% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct-subject-columns.R
\name{distinct_subject_columns}
\alias{distinct_subject_columns}
\title{Determine columns with only one unique per subject.}
\usage{
distinct_subject_columns(.data, .subject_col)
}
\arguments{
\item{.data}{Data.frame}

\item{.subject_col}{String. column name in data for unique subject identifier.}
}
\description{
Return a data.frame with all columns that have one unique value per subject.
}
