% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute-data-diffs.R
\name{execute_data_diffs}
\alias{execute_data_diffs}
\title{Execute Differences Between Data Frames}
\usage{
execute_data_diffs(.base_df, .compare_df, .subject_col, .print_output = TRUE)
}
\arguments{
\item{.base_df}{A data frame that serves as the base for comparison.}

\item{.compare_df}{A data frame to compare against the base data frame.}

\item{.subject_col}{A string representing the column name to be used for subject-based differences.
Set to \code{NULL} for data frames not containing an subject column.}

\item{.print_output}{Logical. Should console output be printed?}
}
\value{
A list with elements \code{diffs}, \code{standard_diffs}, and \code{variable_diffs}
(tibbles summarizing the differences).
}
\description{
This function executes the comparison of two data frames, generates
the differences, and writes them to specified output files.
}
