% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{explain}
\alias{explain}
\title{Explain the Reasoning for a Section of Code}
\usage{
explain(.msg, .expr)
}
\arguments{
\item{.msg}{A character string containing the explanatory message for the code.}

\item{.expr}{An expression representing the R code to be executed.}
}
\value{
Invisible NULL. This function is used for its side effect of printing to the console.
}
\description{
This function prints a formatted message explaining the reasoning for a section of code
and then prints the result of evaluating the provided expression. This function is
intended to be run interactively in the R console to aid in code understanding and debugging.
}
\examples{
\dontrun{
dat <- Theoph
dat$Time[2] <- 0
explain("Subject 1 has duplicate time 0", {
  dat \%>\% filter(Subject == 1, Time == 0)
})
}

}
