% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v-helpers.R
\name{filter_v_subject}
\alias{filter_v_subject}
\title{Filter dataframe by subject column}
\usage{
filter_v_subject(.df, .subject_col = NULL, .subject_filter)
}
\arguments{
\item{.df}{A dataframe that you want to process and view.}

\item{.subject_col}{A character string specifying the subject column. If this column is
present in the dataframe, it will be used to group rows in the datatable.}

\item{.subject_filter}{a string to seach for within the defined \code{.subject_col}}
}
\value{
a filtered dataframe
}
\description{
Filter dataframe by subject column
}
\examples{
\dontrun{
filter_v_subject(df, "USUBJID", "3053-4")
}

}
\keyword{internal}
