% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v-helpers.R
\name{find_df_with_col}
\alias{find_df_with_col}
\title{Search list of dataframes for column}
\usage{
find_df_with_col(.col_name, .df_list)
}
\arguments{
\item{.col_name}{column name}

\item{.df_list}{A dataframe \emph{\strong{or}} list of dataframes that you want to process and view.
Named lists will inherit their names when shown in the app.}
}
\value{
tibble
}
\description{
Search list of dataframes for column
}
\keyword{internal}
