% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-app-bg.R
\name{fmt_bg_file}
\alias{fmt_bg_file}
\title{Format lines of \code{stderr_file} into \code{cli::cli_bullets}}
\usage{
fmt_bg_file(bg_file)
}
\arguments{
\item{bg_file}{file path of background file, such as \code{stderr}}
}
\value{
character vector formatted to be used with \code{cli::cli_bullets}
}
\description{
Format lines of \code{stderr_file} into \code{cli::cli_bullets}
}
\details{
Purpose of this function is to format errors or warnings with use of \code{v()}.
Failure of the background shiny app has somewhat verbose messages, and this
helps to emphasize what actually went wrong.

See \code{?cli::cli_bullets} for how to change formatting
}
\keyword{internal}
