% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-app-bg.R
\name{fmt_bg_msgs}
\alias{fmt_bg_msgs}
\title{Format lines of \code{stderr_file} into \code{cli::cli_bullets}}
\usage{
fmt_bg_msgs(stderr_file)
}
\arguments{
\item{stderr_file}{file path the standard error of the child R process is written to.}
}
\description{
Format lines of \code{stderr_file} into \code{cli::cli_bullets}
}
\keyword{internal}
