% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v-helpers.R
\name{format_v_headers}
\alias{format_v_headers}
\title{Format headers to include column attributes in create_v_datatable (label and class)}
\usage{
format_v_headers(.df)
}
\arguments{
\item{.df}{A dataframe that you want to process and view.}
}
\description{
Format headers to include column attributes in create_v_datatable (label and class)
}
\keyword{internal}
