% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather-data-history.R
\name{gather_data_history}
\alias{gather_data_history}
\title{Gather Data History}
\usage{
gather_data_history(.cur_data, .cur_history, .comment = NULL, .prev_rev = NULL)
}
\arguments{
\item{.cur_data}{A data frame containing the current version of the data.}

\item{.cur_history}{A data frame containing the current history/revision log of the data.}

\item{.comment}{A character string representing a comment or note about the current revision.
Defaults to "No comment" if not provided.}

\item{.prev_rev}{A value representing the previous revision or version number. If not provided,
it won't be logged.}
}
\description{
This function takes the current data, its associated history, and additional
metadata to record a new entry in the data's revision history. It provides
a systematic way to keep track of changes, the author of the change, and more.
}
\keyword{internal}
