% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather-data-labels.R
\name{gather_data_labels}
\alias{gather_data_labels}
\title{View source data domains and column labels}
\usage{
gather_data_labels(.df_list)
}
\arguments{
\item{.df_list}{a named list with each source data domain (ideally use the output of \code{read_src_dir()})}
}
\description{
After loading in your source data, use \code{gather_data_labels()} to view the
description of each domain and column.
}
\keyword{internal}
