% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v-helpers.R
\name{gather_v_cols}
\alias{gather_v_cols}
\title{Determine subject column and which columns can be frozen}
\usage{
gather_v_cols(.df_list, .subject_col = NULL, .id_cols = c("USUBJID", "ID"))
}
\arguments{
\item{.df_list}{A dataframe \emph{\strong{or}} list of dataframes that you want to process and view.
Named lists will inherit their names when shown in the app.}

\item{.subject_col}{A character string specifying the subject column. If this column is
present in the dataframe, it will be used to group rows in the datatable.}

\item{.id_cols}{vector of id columns to search for}
}
\value{
a named list of subject column and available freeze columns
}
\description{
Search for the specified ID columns across list of dataframes, and return the
column that appears the most as the subject column (if not specified).
The freeze columns will be all columns outside of the determined subject column
}
\keyword{internal}
