% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-base-df.R
\name{get_base_df}
\alias{get_base_df}
\title{Get Base Data Frame}
\usage{
get_base_df(.prev_file, .compare_from_svn)
}
\arguments{
\item{.prev_file}{A string indicating the path of the previous file.}

\item{.compare_from_svn}{A boolean flag to determine if the function should compare from svn.}
}
\value{
A list containing two elements: 'base_df' and 'from_svn'.
'base_df' is the dataframe read from the file (or NULL if the file does not exist).
'from_svn' is a boolean indicating whether the file was successfully exported from svn.
}
\description{
This function is used to get the base dataframe from a previous file or from an svn file.
If the \code{.compare_from_svn} flag is set to TRUE, it will try to export the previous file from svn.
If the file does not exist, it will return NULL.
}
\keyword{internal}
