% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-svn-baseline.R
\name{get_svn_baseline}
\alias{get_svn_baseline}
\title{Get Base Data Frame}
\usage{
get_svn_baseline(
  .prev_file,
  .compare_from_svn,
  .reader = read_csv_dots,
  .file_ext = ".csv"
)
}
\arguments{
\item{.prev_file}{A string indicating the path of the previous file.}

\item{.compare_from_svn}{A boolean flag to determine if the function should compare from svn.}

\item{.reader}{A function used to read the file. Defaults to \code{read_csv_dots}.}

\item{.file_ext}{A string indicating the file extension for the temp file. Defaults to \code{".csv"}.}
}
\value{
A list containing:
\itemize{
\item \code{base_df}: the dataframe read from the file (or NULL if the file does not exist).
\item \code{from_svn}: a boolean indicating whether the file was successfully exported from svn.
\item \code{prev_rev}: the SVN revision number (or NA).
\item \code{svn_author}: the last-changed author from SVN (or NA).
\item \code{svn_date}: the last-changed date from SVN (or NA).
}
}
\description{
This function is used to get the base dataframe from a previous file or from an svn file.
If the \code{.compare_from_svn} flag is set to TRUE, it will try to export the previous file from svn.
If the file does not exist, it will return NULL.
}
\keyword{internal}
