% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id-lookup-tracker.R
\name{id_lookup_tracker}
\alias{id_lookup_tracker}
\title{Subject ID Assignment Tracking}
\usage{
id_lookup_tracker(.data, .lookup_path, .subject_col = "USUBJID")
}
\arguments{
\item{.data}{A data.frame to have ID column mutated onto.}

\item{.lookup_path}{A string containing the file path of the ID lookup csv file}

\item{.subject_col}{A character string containing the column name of the subject identifying column in the data}
}
\description{
Upon first use, an ID column will be mutated to the given data set according
to the specified subject column. In addition, an ID lookup csv file will be
created containing the assigned ID for each subject.

For each future use, ID's will be assigned according to the lookup csv file. If
new subjects are added to the data, the lookup file will be modified to include
the new subjects to assign them each a unique identifier.
}
\examples{
Theoph2 <-
  Theoph \%>\%
  dplyr::mutate(USUBJID = paste0("Subject-", Subject)) \%>\%
  dplyr::select(-Subject)

df_with_id <-
  id_lookup_tracker(
    Theoph2,
    .lookup_path = paste0(tempdir(), ".csv"),
    .subject_col = "USUBJID"
  )

}
