% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify-subject-cols.R
\name{identify_subject_cols}
\alias{identify_subject_cols}
\title{Identify columns that are constant within subjects}
\usage{
identify_subject_cols(.df, .subject_col)
}
\arguments{
\item{.df}{A data frame to analyze.}

\item{.subject_col}{A string specifying the name of the subject variable.}
}
\value{
A character vector of column names that are constant within each subject.
}
\description{
This function identifies columns in a data frame that are constant within each subject defined by a specified subject variable.
}
\examples{
df <- data.frame(
  subject = c(1, 1, 2, 2),
  x = c(5, 5, 5, 5),
  y = c(1, 2, 1, 2),
  z = c("a", "a", "a", "a")
)
identify_subject_cols(df, "subject")
# Returns c("x", "z")

}
\keyword{internal}
