% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-v-caption.R
\name{make_v_caption}
\alias{make_v_caption}
\title{Make the caption for \code{create_v_datatable()}, returning the number of subjects}
\usage{
make_v_caption(.df, .subject_col, font_size = 12)
}
\arguments{
\item{.df}{A dataframe that you want to process and view.}

\item{.subject_col}{A character string specifying the subject column. If this column is
present in the dataframe, it will be used to group rows in the datatable.}

\item{font_size}{font size of the caption}
}
\description{
Make the caption for \code{create_v_datatable()}, returning the number of subjects
}
\keyword{internal}
