% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-summary.R
\name{nm_summary}
\alias{nm_summary}
\title{Provide diagnostic summary of an NMTRAN dataset}
\usage{
nm_summary(.data, .spec)
}
\arguments{
\item{.data}{a data frame}

\item{.spec}{a yspec object}
}
\description{
Provide diagnostic summary of an NMTRAN dataset
}
