% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-validate.R
\name{nm_validate}
\alias{nm_validate}
\title{Validate an NMTRAN dataset}
\usage{
nm_validate(.data, .spec, .error_on_fail = TRUE)
}
\arguments{
\item{.data}{a data frame}

\item{.spec}{a yspec object}

\item{.error_on_fail}{if \code{TRUE}, an R error is executed upon failures}
}
\description{
Validate an NMTRAN dataset
}
\examples{

nm_spec <- yspec::ys_load(system.file("derived", "pk.yml", package = "nmvalidate"))

nm <- readr::read_csv(system.file("derived", "pk.csv", package = "nmvalidate"), na = ".")

nm_validate(nm, nm_spec)

}
