% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-write.R
\name{nm_write}
\alias{nm_write}
\title{Write an NMTRAN data set and corresponding meta data
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
nm_write(
  .data,
  .spec,
  .file,
  .prev_file = NULL,
  .compare_from_svn = TRUE,
  .return_base_compare = FALSE
)
}
\arguments{
\item{.data}{a data frame}

\item{.spec}{a yspec object}

\item{.file}{csv file name to write out to (including path)}

\item{.prev_file}{csv file name of previous version (defaults to .file)}

\item{.compare_from_svn}{logical. Should the data comparison be done on the latest svn version? (If not, local version is used)}

\item{.return_base_compare}{logical. Should the two current and previous versions of the datasets be returned?}
}
\description{
\code{nm_write} is deprecated as of mrgda version 0.7.0. Please use \code{write_derived} instead.

This function will take a nonmem ready data frame in R and write it out to a nonmem ready csv.
It also creates a meta data folder, storing the xpt file along with other useful information.
}
\examples{
\dontrun{
nm_spec <- yspec::ys_load(system.file("derived", "pk.yml", package = "mrgda"))
nm <- readr::read_csv(system.file("derived", "pk.csv", package = "mrgda"), na = ".")
nm_write(.data = nm, .spec = nm_spec, .file = "data/derived/pk.csv")
}
}
