% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyNum2.R
\name{prettyNum2}
\alias{prettyNum2}
\title{Round numeric value with significant digits}
\usage{
prettyNum2(.x, .digits = 3)
}
\arguments{
\item{.x}{numeric value}

\item{.digits}{number of digits to round numeric value to}
}
\value{
numeric value
}
\description{
Round numeric value with significant digits
}
\keyword{internal}
