% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-src-lists.R
\name{print.src_list_comparison}
\alias{print.src_list_comparison}
\title{Print method for src_list_comparison}
\usage{
\method{print}{src_list_comparison}(x, ...)
}
\arguments{
\item{x}{A src_list_comparison object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for src_list_comparison
}
