% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-src-list.R
\name{query_src_list}
\alias{query_src_list}
\title{Query a source list.}
\usage{
query_src_list(.src_list, .string)
}
\arguments{
\item{.src_list}{A source list created using read_src_dir.}

\item{.string}{string to search for (not case sensitive)}
}
\description{
Use this to search for a character string across every element in a source data list including
the name, label and contents.
}
\examples{
path <- system.file("example-sdtm", package = "mrgda")

src_list <- read_src_dir(path, .file_types = "xpt")

# Summarize source directory
query_src_list(src_list, .string = "RACE")

}
