% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-app-bg.R
\name{random_dynamic_port}
\alias{random_dynamic_port}
\title{Assign dynamic port}
\usage{
random_dynamic_port(host, n = 20)
}
\arguments{
\item{host}{Host to serve the Shiny app on}

\item{n}{Number of ports to try before giving up}
}
\value{
A port that is available to listen on
}
\description{
Assign dynamic port
}
\keyword{internal}
