% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv-utils.R
\name{read_csv_mrgda}
\alias{read_csv_mrgda}
\title{Read CSV File}
\usage{
read_csv_mrgda(
  file,
  sep = ",",
  quote = FALSE,
  na.strings = ".",
  integer64 = "double",
  data.table = FALSE,
  ...
)
}
\arguments{
\item{file}{A character string specifying the path to the CSV file to be read.}

\item{sep}{Separator between columns, defaults to ","}

\item{quote}{Set to handle fields starting with a double quote, defaults to FALSE}

\item{na.strings}{A character vector of strings to be interpreted as NA values, defaults to "."}

\item{integer64}{Reads columns detected as containing integers larger than 2^31, default to "double"}

\item{data.table}{Boolean, set to TRUE to return data.table, set to FALSE to return data.frame}

\item{...}{arguments passed to \link[data.table]{fread}}
}
\value{
A data frame with the contents of the CSV file, with "." values replaced by \code{NA}.
}
\description{
This function reads a CSV file formatted for NONMEM (or similar) where missing values
are represented by a period (".") and converts these to \code{NA}. It utilizes the
\code{data.table::fread} function for fast reading and converts the result to a data frame.
}
