% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv-utils.R
\name{read_nm_csv}
\alias{read_nm_csv}
\title{Read csv file}
\usage{
read_nm_csv(.file)
}
\arguments{
\item{.file}{file path to csv file}
}
\description{
Reads csv files and formats the data to replace any "." with
NA values.
}
