% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-src-dir.R
\name{read_src_dir}
\alias{read_src_dir}
\title{Read in a source data directory (e.g. a folder filled with SDTM domains).}
\usage{
read_src_dir(.path, .file_types = "detect")
}
\arguments{
\item{.path}{Full path to the source data directory.}

\item{.file_types}{Type of files being read in (e.g. 'xpt').
Setting to 'detect' will determine file type based on the most occurring file type in .path.}
}
\description{
This function will take a nonmem ready data frame in R and write it out to a nonmem ready csv.
It also creates a meta data folder, storing the xpt file along with other useful information.
}
\examples{
path <- system.file("example-sdtm", package = "mrgda")
src_list <- read_src_dir(.path = path)
}
