% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-all-sdtm-checks.R
\name{run_all_sdtm_checks}
\alias{run_all_sdtm_checks}
\title{Run all SDTM checks}
\usage{
run_all_sdtm_checks(.src_list, .subject_col = "USUBJID")
}
\arguments{
\item{.src_list}{List of all soure domain data.frames}

\item{.subject_col}{Column name of subject identifier (character)}
}
\description{
Run all SDTM checks
}
