% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm-check-duplicates.R
\name{sdtm_check_duplicates}
\alias{sdtm_check_duplicates}
\title{Check if duplicates exist in SDTM domain}
\usage{
sdtm_check_duplicates(
  .domain_df,
  .domain_name,
  .cols_check = NULL,
  .subject_col = "USUBJID",
  .domain_filter = NULL
)
}
\arguments{
\item{.domain_df}{source data (data.frame)}

\item{.domain_name}{Name of source domain (character)}

\item{.cols_check}{Columns to search duplicates across (ie. c("USUBJID", "PCDTC"))}

\item{.subject_col}{Column name of subject identifier (character)}

\item{.domain_filter}{Code to use inside a dplyr::filter on the source data.frame before duplicates check}
}
\description{
Check if duplicates exist in SDTM domain
}
\examples{
\dontrun{
sdtm_check_duplicates(.domain_df = src_list$dm, .domain_name = "dm", .cols_check = NULL, .subject_col = "USUBJID", .domain_filter = NULL)
sdtm_check_duplicates(.domain_df = src_list$ex, .domain_name = "ex", .cols_check = c("USUBJID", "EXSTDTC", "EXTRT"))
sdtm_check_duplicates(.domain_df = src_list$pc, .domain_name = "pc", .domain_filter = "PCSTAT != 'NOT DONE'")
}
}
