% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm-check-missing-datetime.R
\name{sdtm_check_missing_datetime}
\alias{sdtm_check_missing_datetime}
\title{Check if missing datetimes in SDTM domain}
\usage{
sdtm_check_missing_datetime(
  .domain_df,
  .domain_name,
  .cols_check = NULL,
  .subject_col = "USUBJID",
  .domain_filter = NULL
)
}
\arguments{
\item{.domain_df}{source data (data.frame)}

\item{.domain_name}{Name of source domain (character)}

\item{.subject_col}{Column name of subject identifier (character)}

\item{.domain_filter}{Code to use inside a dplyr::filter on the source data.frame before duplicates check}

\item{.col_check}{Column to search for missing date/times (ie. c("EXSTDTC"))}
}
\description{
Check if missing datetimes in SDTM domain
}
\examples{
\dontrun{
sdtm_check_missing_datetime(.domain_df = src_list$pc, .domain_name = "pc", .domain_filter = "PCSTAT != 'NOT DONE'")
sdtm_check_missing_datetime(.domain_df = src_list$ex, .domain_name = "ex")
}
}
