% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm-check-missing-subjects.R
\name{sdtm_check_missing_subjects}
\alias{sdtm_check_missing_subjects}
\title{Check if subjects in one domain are missing in others}
\usage{
sdtm_check_missing_subjects(
  .domain_df1,
  .domain_df2,
  .domain_name1,
  .domain_name2,
  .subject_col = "USUBJID",
  .domain_filter1 = NULL,
  .domain_filter2 = NULL
)
}
\arguments{
\item{.domain_df1}{First source data domain to compare (data.frame)}

\item{.domain_df2}{Second source data domain to compare (data.frame)}

\item{.domain_name1}{Name of first source domain (character)}

\item{.domain_name2}{Name of second source domain (character)}

\item{.subject_col}{Column name of subject identifier (character, default = "USUBJID")}

\item{.domain_filter1}{Code to use inside a dplyr::filter on the first source data.frame (optional)}

\item{.domain_filter2}{Code to use inside a dplyr::filter on the second source data.frame (optional)}
}
\description{
Check if subjects in one domain are missing in others
}
