% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signif2.R
\name{signif2}
\alias{signif2}
\title{Round numeric value with significant digits}
\usage{
signif2(.x, .digits = 4)
}
\arguments{
\item{.x}{numeric value}

\item{.digits}{number of digits to round numeric columns to.}
}
\value{
numeric value
}
\description{
Round numeric value with significant digits
}
\keyword{internal}
