% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silence-console-output.R
\name{silence_console_output}
\alias{silence_console_output}
\title{Silence All Console Outputs from an Expression}
\usage{
silence_console_output(.expr)
}
\arguments{
\item{.expr}{An expression to be evaluated.}
}
\value{
The result of the expression \code{.expr} if it completes successfully.
}
\description{
Executes an expression while silencing all messages, warnings, errors, and console outputs.
This includes suppressing messages, warnings, catching errors without printing them to the console,
and suppressing all printed output.
}
\keyword{internal}
