% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src-viz.R
\name{src_viz}
\alias{src_viz}
\title{src_viz function}
\usage{
src_viz(.src_list)
}
\arguments{
\item{.src_list}{A named list of data frames to be visualized.}
}
\value{
None. The function launches a Shiny app.
}
\description{
This function launches a Shiny app that allows a user to visualize a list of data frames.
The user can select a data frame from the list to view, filter it based on the 'USUBJID' column,
and color the rows alternately based on the values in the 'USUBJID' column.
The 'label' attribute of each column, if present, is displayed under the column name.
}
\examples{
read_src_dir(here::here("inst", "example-sdtm")) \%>\% src_viz()
}
