% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-src.R
\name{v}
\alias{v}
\title{create_datatable function}
\usage{
v(.df, .subject_cols = NULL, .freeze_cols = NULL, .group_id = FALSE)
}
\arguments{
\item{.df}{A dataframe that you want to process and view.}

\item{.subject_cols}{A character vector specifying the subject columns. If these columns are
present in the dataframe, it will be used to group rows in the datatable.}

\item{.freeze_cols}{A character vector specifying columns to freeze when scrolling horizontally.
\code{.subject_cols} will automatically be appended to this list (i.e. \code{.subject_cols} are always frozen).}

\item{.group_id}{Logical (\code{TRUE}/\code{FALSE}). If \code{TRUE}, add breaks between each unique \code{.subject_cols} grouping.}
}
\value{
An instance of DT::datatable class with the processed dataframe.
}
\description{
This function modifies the input dataframe by converting columns with fewer than 20 unique values to factors.
It adds labels to the column names and sets up a datatable with a range of options. If the specified
subject column is present, it is used to group rows in the table.
}
\details{
If \code{.subject_cols} is \code{NULL}, the column names \code{"USUBJID"} and \code{"ID"} will be searched and set if present.
If \code{.subject_cols} are present in the dataframe, they will be frozen and alternate in color.

Columns with less than 20 unique values are converted to factors. This can be helpful for filtering, which
will show up as the shiny equivalent of \code{selectInput}, as opposed to the default \code{textInput} for character columns.

Columns with label attributes will be appended to column headers as a new line.

You can \emph{drag and drop} columns to move them around in the table. If a column is dragged to a frozen column's location,
the new column will be frozen instead.
}
