% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v.R
\name{v_global_server}
\alias{v_global_server}
\title{Handle all global UIs}
\usage{
v_global_server(
  .df_list,
  .subject_col = NULL,
  .freeze_cols = NULL,
  input,
  output,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{.df_list}{A dataframe \emph{\strong{or}} list of dataframes that you want to process and view.
Named lists will inherit their names when shown in the app.}

\item{.subject_col}{A character string specifying the subject column. If this column is
present in the dataframe, it will be used to group rows in the datatable.}

\item{.freeze_cols}{A character vector specifying columns to freeze when scrolling horizontally.
\code{.subject_col} will automatically be appended to this list (i.e. \code{.subject_col} is always frozen).}

\item{input}{named list of shiny inputs}

\item{output}{named list of shiny outputs}

\item{session}{main shiny session}
}
\description{
Handle all global UIs
}
\keyword{internal}
