% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v.R
\name{v_global_ui}
\alias{v_global_ui}
\title{Construct global header UI for \code{v_shiny_internal}}
\usage{
v_global_ui(.df_list, .subject_col, .freeze_cols)
}
\arguments{
\item{.df_list}{A dataframe \emph{\strong{or}} list of dataframes that you want to process and view.
Named lists will inherit their names when shown in the app.}

\item{.subject_col}{A character string specifying the subject column. If this column is
present in the dataframe, it will be used to group rows in the datatable.}

\item{.freeze_cols}{A tibble of freeze columns. One column containing the column names,
and the other with the domains they were found in.}
}
\value{
a shiny UI
}
\description{
Serves as a wrapper for all global UIs
}
\details{
Note that this is not module
}
\keyword{internal}
