% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v.R
\name{v_mod_server}
\alias{v_mod_server}
\title{shiny module server for \code{v_shiny_internal}}
\usage{
v_mod_server(
  .id,
  .df,
  .subject_col = NULL,
  .freeze_cols = NULL,
  .digits = 3,
  .subject_filter,
  .dt_options
)
}
\arguments{
\item{.id}{shiny module id. Character string}

\item{.df}{A dataframe that you want to process and view.}

\item{.subject_col}{A character string specifying the subject column. If this column is
present in the dataframe, it will be used to group rows in the datatable.}

\item{.freeze_cols}{reactive expression containing columns to freeze for the particular dataset}

\item{.digits}{number of digits to round numeric columns to. Set to \code{NULL} to prevent rounding.}

\item{.subject_filter}{reactive expression pointing to global filter}

\item{.dt_options}{list of options for formatting the \code{DT::datatable}}
}
\value{
a shiny module server object
}
\description{
shiny module server for \code{v_shiny_internal}
}
\keyword{internal}
