% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v.R
\name{v_mod_ui}
\alias{v_mod_ui}
\title{shiny module UI for \code{v_shiny_internal}}
\usage{
v_mod_ui(id)
}
\arguments{
\item{id}{shiny module id. Character string}
}
\value{
a shiny module UI object
}
\description{
shiny module UI for \code{v_shiny_internal}
}
\keyword{internal}
