% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v.R
\name{v_ui}
\alias{v_ui}
\title{shiny module UI for \code{v_shiny_internal}}
\usage{
v_ui(.id, .df_list, .subject_col)
}
\arguments{
\item{.id}{shiny module id. Character string}

\item{.subject_col}{A character string specifying the subject column. If this column is
present in the dataframe, it will be used to group rows in the datatable.}
}
\value{
a shiny module UI object
}
\description{
shiny module UI for \code{v_shiny_internal}
}
\keyword{internal}
