% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-mrgda-flags.R
\name{view_mrgda_flags}
\alias{view_mrgda_flags}
\title{View mrgda flags found in spec
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
\usage{
view_mrgda_flags(.data, .spec, .view = TRUE)
}
\arguments{
\item{.data}{a data frame}

\item{.spec}{a yspec object}

\item{.view}{Logical. Should the output be viewed? (default is TRUE)}
}
\description{
Users can manually specify flags in the specification file to describe
columns in their data set. \code{mrgda} also attempts to automatically detect
specific columns commonly found in NONMEM data sets.
}
\details{
Use this function to view a summary of how columns in the data set are
currently being described. The output has the following columns:

\itemize{
\item{Flag name - flag to be assigned in the spec}
\item{Assigned Columns - data column(s) assigned for each flag}
\item{Description - definition of each flag}
}
}
\examples{
\dontrun{
nm_spec <- yspec::ys_load(system.file("derived", "pk.yml", package = "mrgda"))
nm <- readr::read_csv(system.file("derived", "pk.csv", package = "mrgda"), na = ".")
view_mrgda_flags(.data = nm, .spec = nm_spec)
}

}
\keyword{internal}
