% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-sdtm-domains.R
\name{view_sdtm_domains}
\alias{view_sdtm_domains}
\title{View SDTM Domain Abbreviations}
\usage{
view_sdtm_domains()
}
\description{
Helper function to view typical SDTM domains. Both the abbreviation and domain
description is provided.
}
\examples{
view_sdtm_domains()

}
