% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-spec-flags.R
\name{view_spec_flags}
\alias{view_spec_flags}
\title{View mrgda flags found in spec}
\usage{
view_spec_flags(.data, .spec)
}
\arguments{
\item{.data}{a data frame}

\item{.spec}{a yspec object}
}
\description{
Users can manually specify flags in the specification file to describe
columns in their data set. \code{mrgda} also attempts to automatically detect
specific columns commonly found in NONMEM data sets.
}
\details{
Use this function to view a summary of how columns in the data set are
currently being described. The output has the following columns:

\itemize{
\item{Flag name - flag to be assigned in the spec}
\item{Default Column - data column name \code{mrgda} uses to automatically detect column}
\item{Detected Columns - data column(s) assigned for each flag}
\item{Description - definition of each flag}
}
}
\examples{
\dontrun{
nm_spec <- yspec::ys_load(system.file("derived", "pk.yml", package = "mrgda"))
nm <- readr::read_csv(system.file("derived", "pk.csv", package = "mrgda"), na = ".")
view_spec_flags(.data = nm, .spec = nm_spec)
}

}
