% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-src.R
\name{view_src}
\alias{view_src}
\title{view_src function}
\usage{
view_src(.df, .subject_col)
}
\arguments{
\item{.df}{A dataframe that you want to process and view.}

\item{.subject_col}{A character string specifying the subject column. If this column is
present in the dataframe, it will be used to group rows in the datatable.}
}
\value{
An instance of DT::datatable class with the processed dataframe.
It also includes options for datatable setup such as searchHighlight, scrollX, and pageLength set to TRUE, 5
respectively by default. If .subject_col exists in the dataframe, it is used to group rows.
The columns with less than 20 unique values are converted to factors. It also modifies column names to include
their labels, if any. The labels are extracted from column attributes.
}
\description{
This function modifies the input dataframe by converting columns with fewer than 20 unique values to factors.
It adds labels to the column names and sets up a datatable with a range of options. If the specified
subject column is present, it is used to group rows in the table.
}
\keyword{internal}
