% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-src-data-labels.R
\name{view_src_data_labels}
\alias{view_src_data_labels}
\title{View source data domains and column labels}
\usage{
view_src_data_labels(.src_list)
}
\arguments{
\item{.src_list}{a named list with each source data domain (ideally use the output of \code{read_src_dir()})}
}
\description{
After loading in your source data, use \code{view_src_data_labels()} to view the
description of each domain and column.
}
\keyword{internal}
