% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-src-dir-summary.R
\name{view_src_dir_summary}
\alias{view_src_dir_summary}
\title{View summary of source data directory}
\usage{
view_src_dir_summary(.path)
}
\arguments{
\item{.path}{Full path to the source data directory.}
}
\description{
This function takes a path to a source data directory (typically SDTM or ADaM folder),
reads in the basename of each data file and its size, and returns a dataframe
with each domains name, size and description of its contents
}
\examples{
path <- system.file("example-sdtm", package = "mrgda")

# Summarize source directory
view_src_dir(.path = path)

}
