% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv-utils.R
\name{write_csv_mrgda}
\alias{write_csv_mrgda}
\title{Write CSV File}
\usage{
write_csv_mrgda(
  x,
  file,
  sep = ",",
  quote = FALSE,
  row.names = FALSE,
  na = ".",
  ...
)
}
\arguments{
\item{x}{Any list of same length vectors (eg. data.frame and data.table)}

\item{file}{Output file name}

\item{sep}{Separator between columns, default is ","}

\item{quote}{Should character fields, factor filds and column names be surrounded by double quotes? Defaults to FALSE}

\item{row.names}{Boolean. Should row names be written? Defaults to FALSE}

\item{na}{The string to use for missing values in the data, defaults to "."}

\item{...}{arguments passed to \code{\link[data.table:fwrite]{data.table::fwrite()}}}
}
\description{
This function writes a CSV file formatted for NONMEM (or similar) where missing values
are converted to a period ("."). It utilizes the \code{data.table::fwrite} function.
}
