% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-derived.R
\name{write_derived}
\alias{write_derived}
\title{Write a derived data set and corresponding meta data}
\usage{
write_derived(
  .data,
  .spec,
  .file,
  .prev_file = NULL,
  .compare_from_svn = TRUE,
  .return_base_compare = FALSE,
  .execute_diffs = TRUE
)
}
\arguments{
\item{.data}{a data frame}

\item{.spec}{a yspec object}

\item{.file}{csv file name to write out to (including path)}

\item{.prev_file}{csv file name of previous version (defaults to .file)}

\item{.compare_from_svn}{logical. Should the data comparison be done on the latest svn version? (If not, local version is used)}

\item{.return_base_compare}{logical. Should the two current and previous versions of the datasets be returned?}

\item{.execute_diffs}{logical. Should the diffs be executed?}
}
\description{
This function will take a data frame in R and write it out to csv.
It also creates a meta data folder, storing the xpt file along with other useful information.
}
\examples{
\dontrun{
nm_spec <- yspec::ys_load(system.file("derived", "pk.yml", package = "mrgda"))
nm <- readr::read_csv(system.file("derived", "pk.csv", package = "mrgda"), na = ".")
write_derived(.data = nm, .spec = nm_spec, .file = "data/derived/pk.csv")
}
}
