% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv-utils.R
\name{write_nm_csv}
\alias{write_nm_csv}
\title{Write csv file}
\usage{
write_nm_csv(.data, .file)
}
\arguments{
\item{.data}{data.frame}

\item{.file}{file path to write out csv file}
}
\description{
Writes csv files and formats the data to replace any NA with
"." values.
}
