% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-subj-level.R
\name{write_subj_level}
\alias{write_subj_level}
\title{Write subject level rds files}
\usage{
write_subj_level(.data, .subject_identifier, .path)
}
\arguments{
\item{.data}{data.frame containing data}

\item{.subject_identifier}{column name in data for unique subject identifier}

\item{.path}{file path to write rds file out to}
}
\description{
Write out a rds file containing all columns where there is only 1 unique
value for each subject (or grouping variable selected)
}
