% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign-id.R
\name{assign_id}
\alias{assign_id}
\title{Subject ID Assignment Tracking}
\usage{
assign_id(.data, .lookup_file, .subject_col = "USUBJID")
}
\arguments{
\item{.data}{A data.frame to have ID column mutated onto.}

\item{.lookup_file}{A string containing the file path of the ID lookup csv file}

\item{.subject_col}{A character string containing the column name of the subject identifying column in the data}
}
\value{
A data frame with an additional ID column, which contains the numeric IDs for each subject in the subject column.
}
\description{
Upon first use, an ID column will be mutated to the given data set according
to the specified subject column. In addition, an ID lookup csv file will be
created containing the assigned ID for each subject.

For each future use, ID's will be assigned according to the lookup csv file. If
new subjects are added to the data, the lookup file will be modified to include
the new subjects to assign them each a unique identifier.
}
\details{
This function creates a lookup for subject IDs. If an ID column already exists in the data,
or the specified subject column does not exist the function will stop.
If a lookup file does not exist at the specified path, the function
will create a new lookup and write it to the provided path. If a lookup file does exist,
the function will assign ID to the data according to the lookup.

If subjects exist in the data but not in the lookup, they will be added to the lookup
and assigned an unique ID value.
}
\examples{
Theoph2 <-
  Theoph \%>\%
  dplyr::mutate(USUBJID = paste0("Subject-", Subject)) \%>\%
  dplyr::select(-Subject)

df_with_id <-
  assign_id(
    Theoph2,
    .lookup_file = paste0(tempfile(), ".csv"),
    .subject_col = "USUBJID"
  )

}
