% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-id.R
\name{calc_id}
\alias{calc_id}
\title{ID formula}
\usage{
calc_id(.unique_subject_identifier, .start_id = 1)
}
\arguments{
\item{.unique_subject_identifier}{subject identifier}

\item{.start_id}{numerical value to start creating ID's}
}
\description{
ID formula
}
\keyword{internal}
