% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v-helpers.R
\name{check_subject_col}
\alias{check_subject_col}
\title{Search for ID columns across list of dataframes}
\usage{
check_subject_col(.src_list, id_cols = c("ID", "USUBJID"))
}
\arguments{
\item{.src_list}{A named list of data frames. Each data frame represents a domain that should be visualized. The list names should be the domain names.}

\item{id_cols}{vector of id columns to search for}
}
\description{
Search for ID columns across list of dataframes
}
\keyword{internal}
