% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v-helpers.R
\name{create_global_filter}
\alias{create_global_filter}
\title{Create global filter for the v shiny app}
\usage{
create_global_filter(.subject_col)
}
\arguments{
\item{.subject_col}{A character string specifying the subject column. If this column is
present in the dataframe, it will be used to group rows in the datatable.}
}
\value{
a \code{shiny.tag} object
}
\description{
global filter is embedded in the title of a \code{shinydashboard::box}
}
\keyword{internal}
