% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-global-filter.R
\name{create_global_filter}
\alias{create_global_filter}
\title{Create global filter for the src_viz shiny app}
\usage{
create_global_filter(.subject_col)
}
\arguments{
\item{.subject_col}{A string specifying the name of the subject column for global filtering, allowing users to filter all datasets at once.
Defaults to NULL.}
}
\description{
Create global filter for the src_viz shiny app
}
\keyword{internal}
