% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute-data-diffs.R
\name{execute_data_diffs}
\alias{execute_data_diffs}
\title{Execute Differences Between Data Frames}
\usage{
execute_data_diffs(
  .base_df,
  .compare_df,
  .output_dir,
  .id_col = "ID",
  .header = ""
)
}
\arguments{
\item{.base_df}{A data frame that serves as the base for comparison.}

\item{.compare_df}{A data frame to compare against the base data frame.}

\item{.output_dir}{A string representing the directory where the output difference files will be stored.}

\item{.id_col}{A string representing the column name to be used for subject-based differences.}

\item{.header}{A string representing a box header.}
}
\value{
No return value, but this function will write files to the specified output directory
containing the differences between the input data frames.
}
\description{
This function executes the comparison of two data frames, generates
the differences, and writes them to specified output files.
}
\keyword{internal}
