% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute-data-diffs.R
\name{execute_data_diffs}
\alias{execute_data_diffs}
\title{Execute Differences Between Data Frames}
\usage{
execute_data_diffs(
  .base_df,
  .compare_df,
  .subject_col,
  .base_from_svn = FALSE,
  .print_output = TRUE
)
}
\arguments{
\item{.base_df}{A data frame that serves as the base for comparison.}

\item{.compare_df}{A data frame to compare against the base data frame.}

\item{.subject_col}{A string representing the column name to be used for subject-based differences.
Set to \code{NULL} for data frames not containing an subject column.}

\item{.base_from_svn}{Logical. Was base df exported from svn?}

\item{.print_output}{Logical. Should console output be printed?}
}
\value{
No return value, but this function will write files to the specified output directory
containing the differences between the input data frames.
}
\description{
This function executes the comparison of two data frames, generates
the differences, and writes them to specified output files.
}
