% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v-helpers.R
\name{format_v_headers}
\alias{format_v_headers}
\title{Format headers to include column attributes in create_v_datatable (label and class)}
\usage{
format_v_headers(.df, .wrap_labels = TRUE, .show_labels = TRUE)
}
\arguments{
\item{.df}{A dataframe that you want to process and view.}

\item{.wrap_labels}{Logical (T/F). IF \code{TRUE}, wrap labels to 20 characters, and indent new lines}

\item{.show_labels}{Logical (T/F). IF \code{TRUE}, show label attributes under the column name}
}
\description{
Format headers to include column attributes in create_v_datatable (label and class)
}
\keyword{internal}
