% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather-data-history.R
\name{gather_data_history}
\alias{gather_data_history}
\title{Gather Data History}
\usage{
gather_data_history(
  .cur_history,
  .comment = NULL,
  .meta_data_folder,
  .prev_rev
)
}
\arguments{
\item{.cur_history}{A data frame or tibble containing the current data history.}

\item{.comment}{A comment describing the current revision. If not provided, defaults to "No comment".}

\item{.meta_data_folder}{Path to the folder containing the metadata, specifically "diffs.csv".}

\item{.prev_rev}{A character string specifying the previous revision code or number.}
}
\value{
A combined data frame/tibble of the current revision details with the previous data history.
}
\description{
This function consolidates the current data history with previous revisions. It reads the differences in the metadata
from "diffs.csv", filters the required changes and combines this with the current revision details.
}
\details{
This function operates in several steps:
\enumerate{
\item If \code{.comment} is not provided, it defaults to "No comment".
\item Attempts to read "diffs.csv" from the \code{.meta_data_folder}.
\item Filters the differences read from "diffs.csv" that are new, removed, or added.
\item Creates a new tibble \code{.history} for the current revision details.
\item If differences were found, they're combined with the current revision details in \code{.history}.
\item All columns in \code{.history} are coerced to character type and NAs are replaced with empty strings.
\item Combines \code{.history} with \code{.cur_history} to form the final data history.
}
}
\examples{
\dontrun{
history_prev <- tibble::tibble(
  Author = "John Smith",
  Comment = "Initial commit",
  `Previous Revision` = "",
  Datetime = "2023-10-06 09:58:58")
gather_data_history(
  history_prev,
  .comment = "Updated data",
  .meta_data_folder = "path/to/metadata",
  .prev_rev = "54")
}

}
\keyword{internal}
