% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data-version.R
\name{get_data_version}
\alias{get_data_version}
\title{Get Data Version}
\usage{
get_data_version(.endpoint, .lookup_path)
}
\arguments{
\item{.endpoint}{A character string denoting the endpoint to be used for file naming.}

\item{.lookup_path}{The path to the YAML file containing version information.}
}
\value{
A list containing the information of the latest version and
the corresponding file name.
}
\description{
Reads the latest version of data from a YAML file and
constructs the filename for that data version.
}
\details{
Get data version

This function reads a YAML file containing version information about data and
returns the latest version along with the corresponding file name.
}
\examples{
# Make sure the yaml file exists at the specified path before running
get_data_version(
  .endpoint = "pk",
  .lookup_path = system.file("derived/data-version-lookup.yaml", package = "mrgda")
  )

}
