% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-v-caption.R
\name{make_v_caption}
\alias{make_v_caption}
\title{Make the caption for \code{v()}, returning the number of subjects}
\usage{
make_v_caption(.df, font_size)
}
\arguments{
\item{.df}{A dataframe that you want to process and view.}

\item{font_size}{font size of the caption}
}
\description{
Make the caption for \code{v()}, returning the number of subjects
}
\keyword{internal}
