% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgda-package.R
\docType{package}
\name{mrgda-package}
\alias{mrgda}
\alias{mrgda-package}
\title{mrgda: Tools for Data Assembly}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Run various tests on an NMTRAN dataset to validate it prior to modeling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://metrumresearchgroup.github.io/mrgda/}
  \item \url{https://github.com/metrumresearchgroup/mrgda}
  \item Report bugs at \url{https://github.com/metrumresearchgroup/mrgda/issues}
}

}
\author{
\strong{Maintainer}: Eric Anderson \email{andersone@metrumrg.com}

Authors:
\itemize{
  \item Grace O'Brien \email{graceo@metrumrg.com}
  \item Michael McDermott \email{michaelm@metrumrg.com}
  \item Kyle Barrett \email{barrettk@metrumrg.com}
}

}
\keyword{internal}
