% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_egfr.R
\name{mutate_egfr}
\alias{mutate_egfr}
\title{Calculate EGFR}
\usage{
mutate_egfr(.df, .age, .wt, .serum_creatinine, .sex, .female_value)
}
\arguments{
\item{.df}{data.frame to add EGFR column to}

\item{.age}{age in years}

\item{.wt}{weight in kilograms}

\item{.serum_creatinine}{serum creatinine mg/dL}

\item{.sex}{numeric sex}

\item{.female_value}{numeric value referencing female subjects}
}
\description{
Mutate estimated glomerular filtration rate to a data.frame. The new column
will be named \code{EGFR}.
}
\details{
https://www.kidney.org/content/ckd-epi-creatinine-equation-2021
}
\examples{

test_df <- dplyr::tibble(
  AGE = c(45, 82, 73),
  WT = c(64, 23, 92),
  SC = c(1.02, 1.04, 1.98),
  SEX = c(1, 2, 1),
  RACE = c(1, 2, 3))

test_df \%>\%
mutate_egfr(
 .age = AGE,
 .wt = WT,
 .serum_creatinine = SC,
 .sex = SEX,
 .female_value = 1
 )

}
