% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-id.R
\name{mutate_id}
\alias{mutate_id}
\title{Mutate ID}
\usage{
mutate_id(.df, .unique_subject_identifier, .start_id = 1)
}
\arguments{
\item{.df}{data.frame to add ID column to}

\item{.unique_subject_identifier}{subject identifier}

\item{.start_id}{numerical value for first ID value}
}
\description{
Mutate ID column to a data.frame. The new column will be named \code{ID}.
}
\examples{

test_df <- dplyr::tibble(
  USUBJID = c("rand-stud-001", "rand-stud-002", "rand-stud-003"),
  WT = c(64, 23, 92),
  SEX = c(1, 2, 1))

test_df \%>\%
mutate_id(
 .unique_subject_identifier = USUBJID,
 .start_id = 1
 )

}
