% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-summary.R
\name{nm_summary}
\alias{nm_summary}
\title{Provide diagnostic summary of an NMTRAN dataset}
\usage{
nm_summary(.data, .spec)
}
\arguments{
\item{.data}{a data frame}

\item{.spec}{a yspec object}
}
\description{
Provide diagnostic summary of an NMTRAN dataset
}
\examples{

nm_spec <- yspec::ys_load(system.file("derived", "pk.yml", package = "nmvalidate"))

nm <- readr::read_csv(system.file("derived", "pk.csv", package = "nmvalidate"), na = ".")

nm_summary(nm, nm_spec)

}
