% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-summary.R
\name{nm_summary}
\alias{nm_summary}
\title{Provide diagnostic summary of an NMTRAN dataset}
\usage{
nm_summary(.data, .spec, .study_compare = TRUE, .return_output = FALSE)
}
\arguments{
\item{.data}{a data frame}

\item{.spec}{a yspec object}

\item{.study_compare}{Should the figures and tabled be returned?}
}
\description{
This function is intended to provide a high level overview of variables
derived during a data assembly. The output of this function will provide
the user with a pdf file containing a series of tables to help the user
better understand distributions within the derived data set.
}
\examples{

nm_spec <- yspec::ys_load(system.file("derived", "pk.yml", package = "mrgda"))

nm <- readr::read_csv(system.file("derived", "pk.csv", package = "mrgda"), na = ".")

nm_summary(.data = nm, .spec = nm_spec)

}
