% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-validate.R
\name{nm_validate}
\alias{nm_validate}
\title{Validate an NMTRAN dataset}
\usage{
nm_validate(.data, .spec, .error_on_fail = TRUE)
}
\arguments{
\item{.data}{a data frame}

\item{.spec}{a yspec object}

\item{.error_on_fail}{if \code{TRUE}, an R error is executed upon failures}
}
\description{
This function is intended to perform a series of checks on a derived data
set by leveraging the data specification used during assembly. If a check
fails, the function will output a description of the problem and code to
use for debugging.
}
\examples{

nm_spec <- yspec::ys_load(system.file("derived", "pk.yml", package = "mrgda"))

nm <- readr::read_csv(system.file("derived", "pk.csv", package = "mrgda"), na = ".")

nm_validate(nm, nm_spec)

}
