% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-validate.R
\name{nm_validate}
\alias{nm_validate}
\title{Validate an NMTRAN dataset}
\usage{
nm_validate(.data, .spec, .error_on_fail = TRUE, .test_omit = NULL)
}
\arguments{
\item{.data}{a data frame}

\item{.spec}{a yspec object}

\item{.error_on_fail}{if \code{TRUE}, an R error is executed upon failures}

\item{.test_omit}{Vector of test names to exclude.}
}
\description{
This function is intended to perform a series of checks on a derived data
set by leveraging the data specification used during assembly. If a check
fails, the function will output a description of the problem and code to
use for debugging.
}
\details{
The following checks are run on the given NONMEM data set:
\itemize{
\item{Are duplicate records present? (based on EVID, DVID and any primary_keys flags)}
\item{Does each subject have only one distinct value for each baseline covariates?}
\item{Are any subjects missing covariate values?}
\item{Do any non-finite TIME values exist?}
\item{Is MDV derived appropriately?}
\item{Are all NUM values unique?}
\item{Is AMT calculated correctly using RATE and DUR?}
}
}
\examples{

nm_spec <- yspec::ys_load(system.file("derived", "pk.yml", package = "mrgda"))

nm <- readr::read_csv(system.file("derived", "pk.csv", package = "mrgda"), na = ".")

nm_validate(.data = nm,
            .spec = nm_spec,
            .error_on_fail = FALSE,
            .test_omit = c('MDV not set to 1 when DV is NA'))

}
