% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-write.R
\name{nm_write}
\alias{nm_write}
\title{Write an NMTRAN data set and corresponding meta data}
\usage{
nm_write(.data, .spec, .file, .script_path)
}
\arguments{
\item{.data}{a data frame}

\item{.spec}{a yspec object}

\item{.file}{csv file name to write out to (including path)}

\item{.script_path}{file path of data assembly script}
}
\description{
This function will take a nonmem ready data frame in R and write it out to a nonmem ready csv.
It also creates a meta data folder, storing the xpt file along with other useful information.
}
\examples{
\dontrun{
nm_spec <- yspec::ys_load(system.file("derived", "pk.yml", package = "mrgda"))
nm <- readr::read_csv(system.file("derived", "pk.csv", package = "mrgda"), na = ".")
nm_write(.data = nm, .spec = nm_spec, .file = "data/derived/pk.csv", .script_path = "script/data-assembly.Rmd")
}
}
