% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmvalidate-package.R
\docType{package}
\name{nmvalidate-package}
\alias{nmvalidate}
\alias{nmvalidate-package}
\title{nmvalidate: Validate NMTRAN datasets}
\description{
Run various tests on an NMTRAN dataset to validate it prior to modeling.
}
\author{
\strong{Maintainer}: Eric Anderson \email{andersone@metrumrg.com}

Authors:
\itemize{
  \item Grace O'Brien \email{graceo@metrumrg.com}
}

}
\keyword{internal}
