% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-src-dir.R
\name{query_src_dir}
\alias{query_src_dir}
\title{Source data query}
\usage{
query_src_dir(.src_directory, .string)
}
\arguments{
\item{.src_directory}{file path to the source data directory}

\item{.string}{string to search for}
}
\description{
Use this to search for a character string across every source data domain including
the name, label and contents of each domain.
}
\examples{
path <- system.file("example-sdtm", package = "mrgda")

# Summarize source directory
query_src_dir(.src_directory = path, .string = "RACE")

}
