% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-src-list.R
\name{query_src_list}
\alias{query_src_list}
\title{Query source data domains for a topic}
\usage{
query_src_list(.src_list, .string, .ignore_case = TRUE)
}
\arguments{
\item{.src_list}{A source list created using read_src_dir.}

\item{.string}{string to search for (not case sensitive by default)}

\item{.ignore_case}{logical. Should the search not be case sensitive? Defaults to TRUE}
}
\description{
Search a source list for a string across column names, column labels,
column values so you can quickly locate data of interest.
}
\examples{
path <- system.file("example-sdtm", package = "mrgda")
src_list <- read_src_dir(path, .file_types = "xpt")

query_src_list(src_list, .string = "RACE")
query_src_list(src_list, .string = "F")

}
