% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv-utils.R
\name{read_nm_csv}
\alias{read_nm_csv}
\title{Read CSV File}
\usage{
read_nm_csv(.file)
}
\arguments{
\item{.file}{A character string specifying the path to the CSV file to be read.}
}
\value{
A data frame with the contents of the CSV file, with "." values replaced by \code{NA}.
}
\description{
This function reads a CSV file formatted for NONMEM (or similar) where missing values
are represented by a period (".") and converts these to \code{NA}. It utilizes the
\code{data.table::fread} function for fast reading and converts the result to a data frame.
}
