% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v-helpers.R
\name{setup_v_list}
\alias{setup_v_list}
\title{Put \code{.df_list} in the right format}
\usage{
setup_v_list(.df_list)
}
\arguments{
\item{.df_list}{A dataframe \emph{\strong{or}} list of dataframes that you want to process and view.}
}
\value{
list of dataframes
}
\description{
Put \code{.df_list} in the right format
}
\keyword{internal}
