% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src-viz.R
\name{src_viz}
\alias{src_viz}
\title{Source Visualizer Shiny App}
\usage{
src_viz(.src_list, .subject_col = "USUBJID")
}
\arguments{
\item{.src_list}{A named list of data frames. Each data frame represents a domain that should be visualized. The list names should be the domain names.}

\item{.subject_col}{A string specifying the name of the subject column for filtering. Defaults to "USUBJID".}
}
\value{
A Shiny App for visualizing source data from various domains.
}
\description{
This function creates a Shiny App for visualizing source data. It receives a list of data frames, each representing a different domain, and
produces a tabbed visualizer where each tab corresponds to a domain. A global filter is provided to filter the data by a subject column.
}
\note{
The Shiny App will be launched in the system's default web browser.

This function assumes the use of the shinydashboard, shiny and purrr packages, and they should be installed and loaded.
}
\examples{
\dontrun{
src_viz(read_src_dir(system.file("example-sdtm", package = "mrgda")), .subject_col = "USUBJID")
}

}
