% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v.R
\name{v}
\alias{v}
\title{Format and view a dataframe or list of dataframes in the viewer}
\usage{
v(.df_list, .subject_col = NULL)
}
\arguments{
\item{.df_list}{A dataframe \emph{\strong{or}} list of dataframes that you want to process and view.
Named lists will inherit their names when shown in the app.}

\item{.subject_col}{A character string specifying the subject column. If this column is
present in the dataframe, it will be used to group rows in the datatable.}
}
\value{
The \code{callr} process (invisibly).
}
\description{
Creates a \code{DT::datatable} object for each dataframe and runs it in a background shiny application in the Rstudio viewer pane.
}
\details{
\strong{For each dataframe in \code{.df_list}:}

If \code{.subject_col} is \code{NULL}, the column names \code{"USUBJID"} and \code{"ID"} will be searched and set if present.
If \code{.subject_col} is present in the dataframe, it will be frozen and alternate in color.

Columns with less than 20 unique values are converted to factors. This can be helpful for filtering, which
will show up as the shiny equivalent of \code{selectInput}, as opposed to the default \code{textInput} for character columns.

Columns with label attributes will be appended to column headers as a new line.

You can \emph{drag and drop} columns to move them around in the table. If a column is dragged to a frozen column's location,
the new column will be frozen instead.
}
\note{
\strong{Notes about the background process:}

By running the app as a background process, users will retain access to their console.
The process will end when one of the following things happen:
\itemize{
\item the Rstudio viewer pane process is stopped
\item another \code{v()} call is executed (only one will run at a time)
\item you manually kill it via the steps below:

\if{html}{\out{<div class="sourceCode">}}\preformatted{result <- v(mtcars)
result$kill()

result
PROCESS 'R', finished.
}\if{html}{\out{</div>}}
}
}
\section{Deprecated}{

This function is deprecated and will be removed in future versions.
}

\examples{
\dontrun{

 df_list <- mrgda::read_src_dir(system.file("example-sdtm", package = "mrgda"))

 # "USUBJID" automatically detected as `.subject_col`

 v(df_list)

 # manually specify `.subject_col`

 v(df_list, .subject_col = "USUBJID")

}

}
\keyword{deprecated}
