% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v.R
\name{v}
\alias{v}
\title{View a formatted dataframe in the viewer}
\usage{
v(.df_list, .subject_col = NULL, .freeze_cols = NULL, .digits = 3)
}
\arguments{
\item{.df_list}{A dataframe \emph{or} list of dataframes that you want to process and view.}

\item{.subject_col}{A character string specifying the subject column. If this column is
present in the dataframe, it will be used to group rows in the datatable.}

\item{.freeze_cols}{A character vector specifying columns to freeze when scrolling horizontally.
\code{.subject_col} will automatically be appended to this list (i.e. \code{.subject_col} is always frozen).}

\item{.digits}{number of digits to round numeric columns to. Set to \code{NULL} to prevent rounding.}
}
\description{
Creates a DT::datatable object and runs it in a background shiny application in the Rstudio viewer pane.
}
\details{
This function modifies the input dataframe by converting columns with fewer than 20 unique values to factors.
It adds labels to the column names and sets up a datatable with a range of options. If the specified
subject column is present, it is used to group rows in the table.

If \code{.subject_col} is \code{NULL}, the column names \code{"USUBJID"} and \code{"ID"} will be searched and set if present.
If \code{.subject_col} is present in the dataframe, it will be frozen and alternate in color.

Columns with less than 20 unique values are converted to factors. This can be helpful for filtering, which
will show up as the shiny equivalent of \code{selectInput}, as opposed to the default \code{textInput} for character columns.

Columns with label attributes will be appended to column headers as a new line.

You can \emph{drag and drop} columns to move them around in the table. If a column is dragged to a frozen column's location,
the new column will be frozen instead.
}
