% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v.R
\name{v_global_server}
\alias{v_global_server}
\title{Handle all global UIs}
\usage{
v_global_server(.df_list, input, output, session = getDefaultReactiveDomain())
}
\arguments{
\item{.df_list}{A dataframe \emph{\strong{or}} list of dataframes that you want to process and view.
Named lists will inherit their names when shown in the app.}

\item{input}{named list of shiny inputs}

\item{output}{named list of shiny outputs}

\item{session}{main shiny session}
}
\value{
a list of \code{reactiveValues}
}
\description{
Handle all global UIs
}
\keyword{internal}
