% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v.R
\name{v_mod_server}
\alias{v_mod_server}
\title{shiny module server for \code{v_shiny_internal}}
\usage{
v_mod_server(id, .subject_col = NULL, global_vars)
}
\arguments{
\item{id}{shiny module id. Character string}

\item{.subject_col}{A character string specifying the subject column. If this column is
present in the dataframe, it will be used to group rows in the datatable.}

\item{global_vars}{reactive list of arguments}
}
\value{
a shiny module server object
}
\description{
shiny module server for \code{v_shiny_internal}
}
\keyword{internal}
