% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v-module.R
\name{v_ui}
\alias{v_ui}
\title{shiny module ui for \code{create_v_datatable}}
\usage{
v_ui(id, .df_list, .subject_col)
}
\arguments{
\item{id}{shiny module id. Character string}

\item{.df_list}{A dataframe \emph{or} list of dataframes that you want to process and view.}

\item{.subject_col}{A character string specifying the subject column. If this column is
present in the dataframe, it will be used to group rows in the datatable.}
}
\description{
shiny module ui for \code{create_v_datatable}
}
\keyword{internal}
