% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v-module.R
\name{v_ui}
\alias{v_ui}
\title{shiny module ui for \code{create_v_datatable}}
\usage{
v_ui(id)
}
\arguments{
\item{id}{shiny module id. Character string}
}
\description{
shiny module ui for \code{create_v_datatable}
}
\keyword{internal}
