% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-sdtm-domains.R
\name{view_sdtm_domains}
\alias{view_sdtm_domains}
\title{View SDTM Domain Abbreviations}
\usage{
view_sdtm_domains(.view = TRUE)
}
\arguments{
\item{.view}{Logical. Should the output be viewed?}
}
\description{
View SDTM Domain Abbreviations
}
